requirements.txt
streamlit>=1.48.1
requests>=2.32.4
numpy>=1.26.0
pandas==2.1.3
SQLAlchemy==2.0.23
alembic==1.12.1
toml
# Image & DICOM
Pillow==10.1.0
opencv-python==4.8.1.78
pydicom==2.4.3
pdf2image==1.17.0
PyPDF2==3.0.1
fpdf2==2.8.3

# AI & ML
torch==2.1.0
torchvision==0.16.0
whisper==1.1.10
neurokit2==0.2.7
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.3
numba==0.61.2
llvmlite==0.44.0

# LLM & API
openai==1.35.0
httpx>=0.27.0,<0.29.0
tenacity==9.1.2
tiktoken==0.11.0

# Text & NLP
transformers==4.45.0  # если используете Hugging Face
sentence-transformers==3.0.0  # опционально

# Audio
SpeechRecognition==3.10.0
pydub==0.25.1
gTTS==2.4.0
pygame==2.5.2

# Web & Backend
Flask==3.0.3
fastapi==0.104.1
uvicorn==0.24.0
starlette==0.27.0
python-multipart==0.0.6

# Database

psycopg2-binary==2.9.10
asyncpg==0.29.0

# Security & Auth
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
cryptography==45.0.6

# Plotting & UI
matplotlib==3.10.5
plotly==6.3.0
seaborn==0.13.2
pydeck==0.9.1
altair==5.5.0

# Data & Formats
openpyxl==3.1.5
python-dateutil==2.8.2
pytz==2025.2
tzdata==2025.2
pyarrow==21.0.0
fsspec==2025.7.0
python-dotenv==1.0.0

# Utils
pillow
pycparser==2.22
cffi==1.17.1
cachetools==6.1.0
rsa==4.9.1
ecdsa==0.19.1
protobuf==6.31.1
distro==1.9.0
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0
certifi==2025.8.3
jinja2==3.1.2
markupsafe==3.0.2
click==8.1.8
itsdangerous==2.2.0
werkzeug==3.1.3
tornado==6.5.2
tqdm==4.67.1
joblib==1.5.1
packaging==25.0
more-itertools==10.7.0
exceptiongroup==1.3.0
sniffio==1.3.1
anyio==3.7.1

# Async & Concurrency
async-timeout==5.0.1
uvloop==0.21.0

# Git
gitdb==4.0.12
smmap==5.0.2
GitPython==3.1.45

# JSON & Schema
jsonschema==4.25.0
typing_extensions==4.14.1
annotated-types==0.7.0
rpds-py==0.27.0

# Pydantic (для современных версий)
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.1.0

# WebSockets & Tools
websockets==12.0
httptools==0.6.4
h11==0.16.0
httpcore==1.0.9
librosa==0.10.1
# Optional: Telegram Bot
python-telegram-bot==22.3
# Добавьте новые зависимости
plotly>=5.0.0
scikit-image>=0.19.0
# Optional: PDF & Fonts
fonttools==4.59.0
contourpy==1.3.2
pyparsing==3.2.3
kiwisolver==1.4.8
lazy_loader==0.4
imageio==2.37.0
tifffile==2025.5.10

# Логирование уже встроено в Python, но можно добавить:
colorlog>=6.7.0  # для цветного логирования (опционально)